A filtered vector indexed as if it were not, and two writes through a -1
From a survey for narrowing and index defects. Its excluded list cleared the places
that looked riskiest -- `flip`'s 400^3 grids go through `std::size_t` throughout,
the `element x kGauss` strides promote, the loaders bound their counts -- which is
what makes the four that remain worth acting on.

**`volumesFromFire` returned a filtered vector that every caller indexes with an
unfiltered index.** A gas space with no geometry was `continue`d past, and
`smoke_view` loops `i < volumes.size()` reading `model.gas[i]` inside it. One skip
shifts every compartment after it onto its neighbour's smoke; skipping the last puts
`volumes[findGas(...)]` out of bounds. `fire::Model::validate` rejects both
conditions, so it was guarded by convention rather than by construction.

It emits a zero-extent placeholder now, and the header states the contract: one
entry per `model.gas`, in the same order, always. **The count alone could not have
caught it** -- a skip plus a spurious push keeps the size and breaks the pairing --
so the test walks the names. And the ferry has no degenerate gas space, so the
placeholder branch was unreachable from the existing fixture and the first version
of the test passed with the placeholder deleted; it knocks the floor area out of the
*first* gas space now, which is the position that shifts everything after it.
Reverting to `continue` fails with the second compartment's 188.4 m2 sitting in the
first's slot.

**`bulkhead_probe` wrote through an unchecked `findCompartment`.** It returns -1 on a
miss and `static_cast<std::size_t>(-1)` is `SIZE_MAX`, so a renamed or missing
`aft_hold_p` stores to a wild address rather than reading a wrong one. This ferry has
been short of an authored compartment before -- her mid wing tanks were never written
-- and `water_probe` guards the identical call. Refused now through the same
`print and return out` idiom the function already uses for a failed thermal solve,
and the second use of the same index further down is guarded too.

**`smoke_view` had the same shape three lines after guarding its siblings**: the two
`findGas` lookups above it are checked and bail, and the `findCompartment` between
them was not.

**And `les::Grid::planArea` formed `n[0] * n[1]` in `int` before widening**, where
`flip.hpp` widens each factor separately for the same quantity. Safe today only
because `gridFor` clamps each axis to 512, and `Field` is a public struct whose
`Grid` a caller fills in directly.

200602 checks, 0 failures. `bulkhead_probe` and `smoke_view` both re-run ok.

+++ engine/sim/les.hpp
@@ -198,7 +198,13 @@ struct Grid {
     // compatible right-hand side if the volume the mean is taken over is the one
     // the cells add up to.
     double volume() const { return cellVolume() * cells(); }
-    double planArea() const { return h[0] * h[1] * static_cast<double>(n[0] * n[1]); }
+    // Each factor widened before the multiply, as `flip.hpp` does for the same
+    // quantity. `n[0] * n[1]` formed the product in `int` first, which is safe only
+    // because `gridFor` clamps each axis to 512 -- and `Field` is a public struct
+    // whose `Grid` a caller can fill in directly.
+    double planArea() const {
+        return h[0] * h[1] * static_cast<double>(n[0]) * static_cast<double>(n[1]);
+    }
     double centre(int axis, int index) const {
         return lo[axis] + (static_cast<double>(index) + 0.5) * h[axis];
     }
